Update filtering code

+++ ddb_ngsflow/variation/variation.py
@@ -207,7 +207,7 @@ def filter_low_quality_variants(job, config, sample, caller, input_vcf):
     for variant in vcf:
         pass_filter = True
         var_info = parse_functions[caller](variant)
-        if int(var_info['Alt_Depth']) < 5:
+        if float(var_info['Alt_Depth']) < 5.0:
             pass_filter = False
         if pass_filter:
             writer.write_record(variant)
